+++ sdk/resourcemanager/datadog/armdatadog/fake/creationsupported_server.go
@@ -15,7 +15,7 @@ import (
 	azfake "github.com/Azure/azure-sdk-for-go/sdk/azcore/fake"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/fake/server"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
-	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/datadog/armdatadog"
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/datadog/armdatadog/v2"
 	"net/http"
 	"net/url"
 	"regexp"
@@ -26,27 +26,19 @@ type CreationSupportedServer struct {
 	// Get is the fake for method CreationSupportedClient.Get
 	// HTTP status codes to indicate success: http.StatusOK
 	Get func(ctx context.Context, datadogOrganizationID string, options *armdatadog.CreationSupportedClientGetOptions) (resp azfake.Responder[armdatadog.CreationSupportedClientGetResponse], errResp azfake.ErrorResponder)
-
-	// NewListPager is the fake for method CreationSupportedClient.NewListPager
-	// HTTP status codes to indicate success: http.StatusOK
-	NewListPager func(datadogOrganizationID string, options *armdatadog.CreationSupportedClientListOptions) (resp azfake.PagerResponder[armdatadog.CreationSupportedClientListResponse])
 }
 
 // NewCreationSupportedServerTransport creates a new instance of CreationSupportedServerTransport with the provided implementation.
 // The returned CreationSupportedServerTransport instance is connected to an instance of armdatadog.CreationSupportedClient via the
 // azcore.ClientOptions.Transporter field in the client's constructor parameters.
 func NewCreationSupportedServerTransport(srv *CreationSupportedServer) *CreationSupportedServerTransport {
-	return &CreationSupportedServerTransport{
-		srv:          srv,
-		newListPager: newTracker[azfake.PagerResponder[armdatadog.CreationSupportedClientListResponse]](),
-	}
+	return &CreationSupportedServerTransport{srv: srv}
 }
 
 // CreationSupportedServerTransport connects instances of armdatadog.CreationSupportedClient to instances of CreationSupportedServer.
 // Don't use this type directly, use NewCreationSupportedServerTransport instead.
 type CreationSupportedServerTransport struct {
-	srv          *CreationSupportedServer
-	newListPager *tracker[azfake.PagerResponder[armdatadog.CreationSupportedClientListResponse]]
+	srv *CreationSupportedServer
 }
 
 // Do implements the policy.Transporter interface for CreationSupportedServerTransport.
@@ -63,8 +55,6 @@ func (c *CreationSupportedServerTransport) Do(req *http.Request) (*http.Response
 	switch method {
 	case "CreationSupportedClient.Get":
 		resp, err = c.dispatchGet(req)
-	case "CreationSupportedClient.NewListPager":
-		resp, err = c.dispatchNewListPager(req)
 	default:
 		err = fmt.Errorf("unhandled API %s", method)
 	}
@@ -102,41 +92,6 @@ func (c *CreationSupportedServerTransport) dispatchGet(req *http.Request) (*http
 	resp, err := server.MarshalResponseAsJSON(respContent, server.GetResponse(respr).CreateResourceSupportedResponse, req)
 	if err != nil {
 		return nil, err
-	}
-	return resp, nil
-}
-
-func (c *CreationSupportedServerTransport) dispatchNewListPager(req *http.Request) (*http.Response, error) {
-	if c.srv.NewListPager == nil {
-		return nil, &nonRetriableError{errors.New("fake for method NewListPager not implemented")}
-	}
-	newListPager := c.newListPager.get(req)
-	if newListPager == nil {
-		const regexStr = `/subscriptions/(?P<subscriptionId>[!#&$-;=?-\[\]_a-zA-Z0-9~%@]+)/providers/Microsoft\.Datadog/subscriptionStatuses`
-		regex := regexp.MustCompile(regexStr)
-		matches := regex.FindStringSubmatch(req.URL.EscapedPath())
-		if matches == nil || len(matches) < 1 {
-			return nil, fmt.Errorf("failed to parse path %s", req.URL.Path)
-		}
-		qp := req.URL.Query()
-		datadogOrganizationIDParam, err := url.QueryUnescape(qp.Get("datadogOrganizationId"))
-		if err != nil {
-			return nil, err
-		}
-		resp := c.srv.NewListPager(datadogOrganizationIDParam, nil)
-		newListPager = &resp
-		c.newListPager.add(req, newListPager)
-	}
-	resp, err := server.PagerResponderNext(newListPager, req)
-	if err != nil {
-		return nil, err
-	}
-	if !contains([]int{http.StatusOK}, resp.StatusCode) {
-		c.newListPager.remove(req)
-		return nil, &nonRetriableError{fmt.Errorf("unexpected status code %d. acceptable values are http.StatusOK", resp.StatusCode)}
-	}
-	if !server.PagerResponderMore(newListPager) {
-		c.newListPager.remove(req)
 	}
 	return resp, nil
 }